# Add bufferItems support & other cross-partition perf improvements