--- a/src/ClientContext.ts
+++ b/src/ClientContext.ts
@@ -1,7 +1,9 @@
+import uuid from "uuid";
 import { PartitionKeyRange } from "./client/Container/PartitionKeyRange";
 import { Resource } from "./client/Resource";
 import { Constants, HTTPMethod, OperationType, ResourceType } from "./common/constants";
 import { getIdFromLink, getPathFromLink, parseLink } from "./common/helper";
+import { logger } from "./common/logger";
 import { StatusCodes, SubStatusCodes } from "./common/statusCodes";
 import { CosmosClientOptions } from "./CosmosClientOptions";
 import { ConnectionPolicy, ConsistencyLevel, DatabaseAccount, PartitionKey } from "./documents";
@@ -10,14 +12,17 @@ import { executePlugins, PluginOn } from "./plugins/Plugin";
 import { FetchFunctionCallback, SqlQuerySpec } from "./queryExecutionContext";
 import { CosmosHeaders } from "./queryExecutionContext/CosmosHeaders";
 import { QueryIterator } from "./queryIterator";
-import { FeedOptions, RequestOptions, Response } from "./request";
 import { ErrorResponse } from "./request";
+import { FeedOptions, RequestOptions, Response } from "./request";
 import { PartitionedQueryExecutionInfo } from "./request/ErrorResponse";
 import { getHeaders } from "./request/request";
 import { RequestContext } from "./request/RequestContext";
 import { request as executeRequest } from "./request/RequestHandler";
 import { SessionContainer } from "./session/sessionContainer";
 import { SessionContext } from "./session/SessionContext";
+
+/** @hidden */
+const log = logger("ClientContext");
 
 /**
  * @hidden
@@ -122,7 +127,7 @@ export class ClientContext {
       plugins: this.cosmosClientOptions.plugins,
       partitionKey
     };
-
+    const requestId = uuid.v4();
     if (query !== undefined) {
       request.method = HTTPMethod.post;
     }
@@ -136,7 +141,11 @@ export class ClientContext {
       }
     }
     this.applySessionToken(request);
+    log("info", "query " + requestId + " started");
+    log("debug", request);
+    const start = Date.now();
     const response = await executeRequest(request);
+    log("info", "query " + requestId + " finished - " + (Date.now() - start) + "ms");
     this.captureSessionToken(undefined, path, OperationType.Query, response.headers);
     return this.processQueryFeedResponse(response, !!query, resultFn);
   }