--- a/src/common/logger.ts
+++ b/src/common/logger.ts
@@ -1,18 +1,31 @@
 import debugLib from "debug";
 
+/** @hidden */
 const cosmosLevelFilter = process.env.COSMOS_LOG_LEVEL || "warn|error";
 
 /** @hidden */
 const cosmosDebug = debugLib("cosmos");
 
 /** @hidden */
-export type logLevel = "silly" | "debug" | "info" | "warn" | "error";
+type logLevel = "silly" | "debug" | "info" | "warn" | "error";
+
+/** @hidden */
+const levelLogger = (namespaceLogger: debugLib.Debugger, level: logLevel) => {
+  return (message: string | { [key: string]: any }) => {
+    if (cosmosLevelFilter.includes(level)) {
+      namespaceLogger("[" + new Date().toISOString() + "][" + level + "]: %o", message);
+    }
+  };
+};
 
 /** @hidden */
 export const logger = (namespace: string) => {
-  return (level: logLevel, message: string | { [key: string]: any }) => {
-    if (cosmosLevelFilter.includes(level)) {
-      cosmosDebug("[" + new Date().toISOString() + "][" + level + "][" + namespace + "]: %o", message);
-    }
+  const namespaceLogger = cosmosDebug.extend(namespace);
+  return {
+    silly: levelLogger(namespaceLogger, "silly"),
+    debug: levelLogger(namespaceLogger, "debug"),
+    info: levelLogger(namespaceLogger, "info"),
+    warn: levelLogger(namespaceLogger, "warn"),
+    error: levelLogger(namespaceLogger, "error")
   };
 };