… package.json
… src/queryIterator.ts
… src/request/FeedOptions.ts
… src/request/RequestHandler.ts
… test/integration/crossPartition.spec.ts
… test/unit/defaultQueryExecutionContext.spec.ts